Cleanup more TODOs

+++ async-io/src/main/java/org/reactivetoolbox/io/uring/struct/raw/SubmitQueueEntry.java
@@ -63,12 +63,8 @@ public class SubmitQueueEntry extends AbstractExternalRawStructure<SubmitQueueEn
         return putShort(ioprio, data);
     }
 
-    //TODO: watch until it will be included into liburing
-//    public SubmitQueueEntry pollEvents(final short data) {
-//        return putShort(poll_events, data);
-//    }
-    public SubmitQueueEntry pollEvents(final int data) {
-        return putInt(poll_events, data);
+    public SubmitQueueEntry pollEvents(final short data) {
+        return putShort(poll_events, data);
     }
 
     public SubmitQueueEntry bufIndex(final short data) {
